# frontend/schema/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OverlayPayload": {
        "description": "Predefined overlay entry returned by ``GET /overlays``.",
        "properties": {
          "name": {
            "description": "Display name of the overlay",
            "title": "Name",
            "type": "string"
          },
          "oid": {
            "description": "Overlay identifier",
            "title": "Oid",
            "type": "string"
          }
        },
        "required": [
          "name",
          "oid"
        ],
        "title": "OverlayPayload",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Volley Overlay Control",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/links": {
      "get": {
        "description": "Return control, overlay, and preview links for the session.",
        "operationId": "get_links_api_v1_links_get",
        "parameters": [
          {
            "description": "Overlay ID",
            "in": "query",
            "name": "oid",
            "required": true,
            "schema": {
              "description": "Overlay ID",
              "title": "Oid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Links",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
    "/api/v1/overlays": {
      "get": {
        "description": "Return predefined overlays available for selection.\n\nSourced exclusively from the ``PREDEFINED_OVERLAYS`` environment\nvariable (also populated via the remote configurator). Entries are\nfiltered by ``allowed_users`` using the caller's identity when user\nauthentication is enabled.",
        "operationId": "get_overlays_api_v1_overlays_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OverlayPayload"
                  },
                  "title": "Response Get Overlays Api V1 Overlays Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Overlays",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
...
    "/api/v1/styles": {
      "get": {
        "description": "Return available overlay styles.",
        "operationId": "get_styles_api_v1_styles_get",
        "parameters": [
          {
            "description": "Overlay ID",
            "in": "query",
            "name": "oid",
            "required": true,
            "schema": {
              "description": "Overlay ID",
              "title": "Oid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Styles",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
    "/api/v1/teams": {
      "get": {
        "description": "Return predefined team names with icon/color data.",
        "operationId": "get_teams_api_v1_teams_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Teams",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
    "/api/v1/themes": {
      "get": {
        "description": "Return available theme definitions.",
        "operationId": "get_themes_api_v1_themes_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Themes",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
    "/create/overlay/{overlay_id}": {
      "get": {
        "operationId": "create_overlay_create_overlay__overlay_id__get_post",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Overlay",
        "tags": [
          "Overlay"
        ]
      },
      "post": {
        "operationId": "create_overlay_create_overlay__overlay_id__get_post",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Overlay",
        "tags": [
          "Overlay"
        ]
      }
    },
    "/delete/overlay/{overlay_id}": {
      "delete": {
        "operationId": "delete_overlay_delete_overlay__overlay_id__delete_get_post",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Overlay",
        "tags": [
          "Overlay"
        ]
      },
      "get": {
        "operationId": "delete_overlay_delete_overlay__overlay_id__delete_get_post",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Overlay",
        "tags": [
          "Overlay"
        ]
      },
      "post": {
        "operationId": "delete_overlay_delete_overlay__overlay_id__delete_get_post",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Overlay",
        "tags": [
          "Overlay"
        ]
      }
    },
...
    "/list/overlay": {
      "get": {
        "description": "Return every overlay id plus its output key.\n\nGated behind ``OVERLAY_MANAGER_PASSWORD`` because the response\ndefeats the capability-URL design of ``/overlay/{output_key}``.\nSee ``AUTHENTICATION.md`` (F-4).",
        "operationId": "list_overlays_list_overlay_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Overlays",
        "tags": [
          "Overlay"
        ]
      }
    },
...
    "/overlay/{overlay_id}": {
      "get": {
        "operationId": "serve_overlay_overlay__overlay_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "overlay_id",
            "required": true,
            "schema": {
              "title": "Overlay Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "style",
            "required": false,
            "schema": {
              "title": "Style",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Serve Overlay",
        "tags": [
          "Overlay"
        ]
      }
    },
...
  }
}